Moved ParseState over to Parser.hpp; proper Parser class to follow.
